Fix a couple of cargo warnings and make the tests/examples actually compile
cursor.rs and http.rs could not be found because they were specified by
filename instead of the full relative path. We don't have to provide
the filename or path at all if it's the same as the name of the target.

And cargo was complaining about the examples/tests inference behaviour
changes between Rust 2015 and 2018. Fixed by explicitly listing all
tests and examples that should be compiled/run in Cargo.toml.